+++ CHANGES.rst
@@ -6,6 +6,7 @@ Version 0.2
 ===========
 
 - Add metadata from netCDF file to returned image.
+- Add option to return data as 1D arrays.
 
 Version 0.1
 ===========